// src/store/modules/upload/action-types.ts
export enum ActionTypes {
  UPLOAD_CONFIGURATIONS = "UPLOAD_CONFIGURATIONS",
  UPLOAD_FILE = "UPLOAD_FILE",
  UPLOADED_FILES = "UPLOADED_FILES",
  UPLOAD_PROJECT_POSTING_FILE = "UPLOAD_PROJECT_POSTING_FILE",
  UPLOADED_PROJECT_POSTING_FILES = "UPLOADED_PROJECT_POSTING_FILES",
  DELETE_FILE = "DELETE_FILE",
  DELETE_PROJECT_POSTING_FILE = "DELETE_PROJECT_POSTING_FILE",
}